+++ programs/onreapp/tests/delete_all_offer_vectors.rs
@@ -43,7 +43,7 @@ fn test_delete_all_multiple_vectors() {
             &token_out,
             None,
             current_time + (i * 1000),
-            i as u64 * 1_000_000,
+            i * 1_000_000,
             5000,
             3600,
         );
